Implement context-generic DelegateBuildConnectionHandshakeMessage
