+++ crates/cosmos/cosmos-chain-components/src/components/cosmos_to_cosmos.rs
@@ -1,5 +1,6 @@
 use cgp_core::prelude::*;
 use hermes_relayer_components::chain::impls::delegate::queries::client_state::QueryAndDecodeClientStateVia;
+use hermes_relayer_components::chain::traits::message_builders::connection_handshake::ConnectionHandshakeMessageBuilderComponent;
 use hermes_relayer_components::chain::traits::message_builders::create_client::CreateClientMessageBuilderComponent;
 use hermes_relayer_components::chain::traits::message_builders::update_client::UpdateClientMessageBuilderComponent;
 use hermes_relayer_components::chain::traits::queries::client_state::{
@@ -10,6 +11,7 @@ use prost_types::Any;
 
 use crate::impls::client::create_client_message::BuildCosmosCreateClientMessage;
 use crate::impls::client::update_client_message::BuildCosmosUpdateClientMessage;
+use crate::impls::connection::connection_handshake_message::BuildCosmosConnectionHandshakeMessage;
 use crate::impls::queries::consensus_state::QueryCosmosConsensusStateFromChainHandle;
 
 pub struct CosmosToCosmosComponents;
@@ -27,5 +29,7 @@ delegate_components! {
             BuildCosmosCreateClientMessage,
         UpdateClientMessageBuilderComponent:
             BuildCosmosUpdateClientMessage,
+        ConnectionHandshakeMessageBuilderComponent:
+            BuildCosmosConnectionHandshakeMessage,
     }
 }
