--- crates/cosmos/cosmos-chain-components/src/components/ibc_client.rs
@@ -1,22 +1,0 @@
-use cgp_core::prelude::*;
-use hermes_relayer_components::chain::traits::message_builders::connection_handshake::ConnectionHandshakeMessageBuilderComponent;
-use hermes_relayer_components::chain::traits::message_builders::update_client::UpdateClientMessageBuilderComponent;
-use hermes_relayer_components::chain::traits::queries::consensus_state_height::ConsensusStateHeightsQuerierComponent;
-
-use crate::impls::client::update_client_message::BuildCosmosUpdateClientMessage;
-use crate::impls::connection::connection_handshake_message::BuildCosmosConnectionHandshakeMessage;
-use crate::impls::queries::consensus_state_height::QueryConsensusStateHeightsFromChainHandle;
-
-pub struct CosmosIbcClientComponents;
-
-delegate_components! {
-    #[mark_component(IsCosmosIbcClientComponents)]
-    CosmosIbcClientComponents {
-        UpdateClientMessageBuilderComponent:
-            BuildCosmosUpdateClientMessage,
-        ConnectionHandshakeMessageBuilderComponent:
-            BuildCosmosConnectionHandshakeMessage,
-        ConsensusStateHeightsQuerierComponent:
-            QueryConsensusStateHeightsFromChainHandle,
-    }
-}
