+++ crates/cosmos/cosmos-chain-components/src/components/mod.rs
@@ -1,5 +1,4 @@
 pub mod client;
 pub mod cosmos_to_cosmos;
 pub mod delegate;
-pub mod ibc_client;
 pub mod transaction;
