+++ crates/relayer/relayer-components/src/chain/impls/delegate/message_builders/connection_handshake.rs
@@ -1,0 +1,81 @@
+use core::marker::PhantomData;
+
+use cgp_core::prelude::HasErrorType;
+use cgp_core::DelegateComponent;
+
+use crate::chain::traits::message_builders::connection_handshake::ConnectionHandshakeMessageBuilder;
+use crate::chain::traits::types::connection::{
+    HasConnectionHandshakePayloadTypes, HasInitConnectionOptionsType,
+};
+use crate::chain::traits::types::ibc::HasIbcChainTypes;
+
+pub struct DelegateBuildConnectionHandshakeMessage<Components>(pub PhantomData<Components>);
+
+impl<Chain, Counterparty, Components, Delegate>
+    ConnectionHandshakeMessageBuilder<Chain, Counterparty>
+    for DelegateBuildConnectionHandshakeMessage<Components>
+where
+    Chain:
+        HasInitConnectionOptionsType<Counterparty> + HasIbcChainTypes<Counterparty> + HasErrorType,
+    Counterparty: HasConnectionHandshakePayloadTypes<Chain> + HasIbcChainTypes<Chain>,
+    Delegate: ConnectionHandshakeMessageBuilder<Chain, Counterparty>,
+    Components: DelegateComponent<Counterparty, Delegate = Delegate>,
+{
+    async fn build_connection_open_init_message(
+        chain: &Chain,
+        client_id: &Chain::ClientId,
+        counterparty_client_id: &Counterparty::ClientId,
+        init_connection_options: &Chain::InitConnectionOptions,
+        counterparty_payload: Counterparty::ConnectionOpenInitPayload,
+    ) -> Result<Chain::Message, Chain::Error> {
+        Delegate::build_connection_open_init_message(
+            chain,
+            client_id,
+            counterparty_client_id,
+            init_connection_options,
+            counterparty_payload,
+        )
+        .await
+    }
+
+    async fn build_connection_open_try_message(
+        chain: &Chain,
+        client_id: &Chain::ClientId,
+        counterparty_client_id: &Counterparty::ClientId,
+        counterparty_connection_id: &Counterparty::ConnectionId,
+        counterparty_payload: Counterparty::ConnectionOpenTryPayload,
+    ) -> Result<Chain::Message, Chain::Error> {
+        Delegate::build_connection_open_try_message(
+            chain,
+            client_id,
+            counterparty_client_id,
+            counterparty_connection_id,
+            counterparty_payload,
+        )
+        .await
+    }
+
+    async fn build_connection_open_ack_message(
+        chain: &Chain,
+        connection_id: &Chain::ConnectionId,
+        counterparty_connection_id: &Counterparty::ConnectionId,
+        counterparty_payload: Counterparty::ConnectionOpenAckPayload,
+    ) -> Result<Chain::Message, Chain::Error> {
+        Delegate::build_connection_open_ack_message(
+            chain,
+            connection_id,
+            counterparty_connection_id,
+            counterparty_payload,
+        )
+        .await
+    }
+
+    async fn build_connection_open_confirm_message(
+        chain: &Chain,
+        connection_id: &Chain::ConnectionId,
+        counterparty_payload: Counterparty::ConnectionOpenConfirmPayload,
+    ) -> Result<Chain::Message, Chain::Error> {
+        Delegate::build_connection_open_confirm_message(chain, connection_id, counterparty_payload)
+            .await
+    }
+}
