+++ crates/relayer/relayer-components/src/chain/impls/delegate/message_builders/mod.rs
@@ -1,2 +1,3 @@
+pub mod connection_handshake;
 pub mod create_client;
 pub mod update_client;
