+++ crates/sovereign/sovereign-relayer/src/exts/cosmos/chain.rs
@@ -1,12 +1,11 @@
 use cgp_core::prelude::*;
 use hermes_cosmos_chain_components::components::delegate::DelegateCosmosChainComponents;
-use hermes_cosmos_relayer::chain::impls::connection_handshake_message::DelegateCosmosConnectionHandshakeBuilder;
 use hermes_cosmos_relayer::contexts::chain::CosmosChain;
 use hermes_relayer_components::chain::impls::delegate::queries::client_state::QueryAndDecodeClientStateVia;
 use hermes_sovereign_chain_components::cosmos::impls::sovereign_to_cosmos::client::create_client_message::BuildCreateSovereignClientMessageOnCosmos;
 use hermes_sovereign_chain_components::cosmos::impls::sovereign_to_cosmos::client::update_client_message::BuildUpdateSovereignClientMessageOnCosmos;
 use hermes_sovereign_chain_components::cosmos::impls::sovereign_to_cosmos::connection_handshake_message::BuildSovereignConnectionHandshakeMessageOnCosmos;
-use hermes_relayer_components::chain::traits::message_builders::connection_handshake::CanBuildConnectionHandshakeMessages;
+use hermes_relayer_components::chain::traits::message_builders::connection_handshake::{CanBuildConnectionHandshakeMessages, ConnectionHandshakeMessageBuilderComponent};
 use hermes_relayer_components::chain::traits::message_builders::create_client::{CanBuildCreateClientMessage, CreateClientMessageBuilderComponent};
 use hermes_relayer_components::chain::traits::message_builders::update_client::{CanBuildUpdateClientMessage, UpdateClientMessageBuilderComponent};
 use hermes_relayer_components::chain::traits::queries::client_state::{AllClientStatesBytesQuerierComponent, CanQueryClientState, ClientStateQuerierComponent};
@@ -33,6 +32,8 @@ delegate_components! {
             BuildUpdateSovereignClientMessageOnCosmos,
         CreateClientMessageBuilderComponent:
             BuildCreateSovereignClientMessageOnCosmos,
+        ConnectionHandshakeMessageBuilderComponent:
+            BuildSovereignConnectionHandshakeMessageOnCosmos,
     }
 }
 
@@ -45,9 +46,3 @@ pub trait CanUseSovereignMethodsOnCosmosChain:
 }
 
 impl CanUseSovereignMethodsOnCosmosChain for CosmosChain {}
-
-delegate_components! {
-    DelegateCosmosConnectionHandshakeBuilder {
-        SovereignChain: BuildSovereignConnectionHandshakeMessageOnCosmos,
-    }
-}
